add Dockerfile. not used yet.
the point of this repo is to simulate a lift-and-shift from render to
aws, specifically ecs. the "c" in ecs means container. so before
adding aws resources, i want to change the render setup to run the app
via docker.

here's how i run it locally:

```
$ docker build --tag r2a .
$ docker run --publish 8080:8080 r2a
```

i'll wire this into render.tf in the next commit (hopefully...).